// src/generated/models/UpdateAitDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AIT  doc
 * AIT documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateAitDto
 */
export interface UpdateAitDto {
    /**
     * Vehicle license plate
     * @type {string}
     * @memberof UpdateAitDto
     */
    placaVeiculo?: string;
    /**
     * Date and time of the infraction
     * @type {Date}
     * @memberof UpdateAitDto
     */
    dataInfracao?: Date;
    /**
     * Description of the infraction
     * @type {string}
     * @memberof UpdateAitDto
     */
    descricao?: string;
    /**
     * Fine amount in decimal
     * @type {number}
     * @memberof UpdateAitDto
     */
    valorMulta?: number;
}

/**
 * Check if a given object implements the UpdateAitDto interface.
 */
export function instanceOfUpdateAitDto(value: object): value is UpdateAitDto {
    return true;
}

export function UpdateAitDtoFromJSON(json: any): UpdateAitDto {
    return UpdateAitDtoFromJSONTyped(json, false);
}

export function UpdateAitDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateAitDto {
    if (json == null) {
        return json;
    }
    return {
        
        'placaVeiculo': json['placa_veiculo'] == null ? undefined : json['placa_veiculo'],
        'dataInfracao': json['data_infracao'] == null ? undefined : (new Date(json['data_infracao'])),
        'descricao': json['descricao'] == null ? undefined : json['descricao'],
        'valorMulta': json['valor_multa'] == null ? undefined : json['valor_multa'],
    };
}

export function UpdateAitDtoToJSON(json: any): UpdateAitDto {
    return UpdateAitDtoToJSONTyped(json, false);
}

export function UpdateAitDtoToJSONTyped(value?: UpdateAitDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'placa_veiculo': value['placaVeiculo'],
        'data_infracao': value['dataInfracao'] == null ? undefined : ((value['dataInfracao']).toISOString()),
        'descricao': value['descricao'],
        'valor_multa': value['valorMulta'],
    };
}